+++ nuxt.config.js
@@ -43,5 +43,15 @@ module.exports = {
         })
       }
     }
+  },
+  build: {
+    vendor: ['vue-i18n']
+  },
+  router: {
+    middleware: 'i18n'
+  },
+  plugins: ['~plugins/i18n.js',],
+  generate: {
+    routes: ['/', '/about', '/fr', '/fr/about']
   }
 }
