+++ package.json
@@ -6,7 +6,8 @@
   "private": true,
   "dependencies": {
     "axios": "^0.16.1",
-    "nuxt": "latest"
+    "nuxt": "latest",
+    "vue-i18n": "^7.0.1"
   },
   "scripts": {
     "dev": "nuxt",
