+++ store/index.js
@@ -1,0 +1,12 @@
+export const state = () => ({
+  locales: ['en', 'fr'],
+  locale: 'en'
+})
+
+export const mutations = {
+  SET_LANG (state, locale) {
+    if (state.locales.indexOf(locale) !== -1) {
+      state.locale = locale
+    }
+  }
+}
